storage 12/12/2024

+++ src/main/java/org/example/storage_service/entity/File.java
@@ -20,7 +20,7 @@ public class File extends AuditableEntity {
     private String filePath;
     private String ratio;
     private String fileSize;
-    private int ownerId;
+    private Long ownerId;
     private String accessType;
     private boolean deleted = false;
     private boolean visible =true;
